requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
graphiti-core[falkordb]>=0.28.2
aiosqlite>=0.20.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
